Added final form validation + fixed bugs
The form is fully functional with all the validation implemented. There
was a minor bug where the zipcode will have issues when a country was
not selected. That has been fixed. CSS styling was also updated. All
that's left is testing all possible outcomes to see if there are any
bugs/issues left.